sound: export tfa98xx header to userspace
