+++ include/uapi/sound/Kbuild
@@ -17,3 +17,4 @@ header-y += audio_effects.h
 header-y += voice_svc.h
 header-y += devdep_params.h
 header-y += msmcal-hwdep.h
+header-y += tfa98xx.h
